# Speed up and cache monospace font searching