Fixes #258

My system: 1300ms of loading fonts --> 1ms of loading fonts

@adder46's system: total startup time 25sec --> 1sec